+++ packages/brick_graphql/lib/src/transformers/model_fields_document_transformer.dart
@@ -110,11 +110,30 @@ class ModelFieldsDocumentTransformer<_Model extends GraphqlModel> {
                   modelDictionary.adapterFor[entry.value.type]!.fieldsToGraphqlRuntimeDefinition,
                 ),
               )
-            : null,
+            : entry.value.subfields.isNotEmpty
+                ? _generateSubFields(entry.value.subfields)
+                : null,
       ));
 
       return nodes;
     });
+  }
+
+  SelectionSetNode _generateSubFields(Set<String> subfields) {
+    return SelectionSetNode(
+      selections: subfields
+          .map<SelectionNode>((subfield) {
+            return FieldNode(
+              name: NameNode(value: subfield),
+              alias: null,
+              arguments: [],
+              directives: [],
+              selectionSet: null,
+            );
+          })
+          .toList()
+          .cast<SelectionNode>(),
+    );
   }
 
   /// Merge the operation headers from [document] and the generated `#document` nodes.